# openapi.yaml
openapi: 3.0.1
info:
  title: Terraware Seed Bank
  description: Local server API for seed banks
  license:
    name: MIT
  version: 0.1-SNAPSHOT
servers:
- url: http://localhost:32109
  description: Generated server url
...
paths:
...
  /api/v1/seedbank/values:
    post:
      tags:
      - SeedBankApp
      summary: List the values of a set of search fields for a set of accessions matching
        certain filter criteria.
      operationId: listFieldValues
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListFieldValuesRequestPayload'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListFieldValuesResponsePayload'
  /api/v1/seedbank/values/all:
    post:
      tags:
      - SeedBankApp
      summary: List the possible values of a set of search fields.
      operationId: listAllFieldValues
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListAllFieldValuesRequestPayload'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAllFieldValuesResponsePayload'
  /api/v1/seedbank/values/species:
    get:
      tags:
      - SeedBankApp
      description: Use /api/v1/species instead.
      operationId: listSpecies
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSpeciesResponsePayload'
      deprecated: true
    post:
      tags:
      - SeedBankApp
      description: Use /api/v1/species instead.
      operationId: createSpecies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpeciesRequestPayload'
        required: true
      responses:
        "200":
          description: Species created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSpeciesResponsePayload'
        "409":
          description: A species with the requested name already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSpeciesResponsePayload'
      deprecated: true
  /api/v1/seedbank/values/species/{id}:
    post:
      tags:
      - SeedBankApp
      summary: Updates an existing species.
      description: "If the species is being renamed and the species name in the payload\
        \ already exists, the existing species replaces the one in the request (thus\
        \ merging the requested species into the one that already had the name) and\
        \ its ID is returned."
      operationId: updateSpecies
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpeciesRequestPayload'
        required: true
      responses:
        "200":
          description: Species updated or merged with an existing species.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateSpeciesResponsePayload'
        "404":
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
...
  /api/v1/species:
    get:
      tags:
      - GISApp
      summary: Lists all known species.
      operationId: speciesList
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesListResponsePayload'
    post:
      tags:
      - GISApp
      summary: Creates a new species.
      operationId: speciesCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeciesRequestPayload'
        required: true
      responses:
        "200":
          description: Species created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesCreateResponsePayload'
        "409":
          description: A species with the requested name already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesCreateResponsePayload'
...
  /api/v1/species/{speciesId}:
    get:
      tags:
      - GISApp
      summary: Gets information about a single species.
      operationId: speciesRead
      parameters:
      - name: speciesId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Species retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesGetResponsePayload'
        "404":
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
    put:
      tags:
      - GISApp
      summary: Updates an existing species.
      operationId: speciesUpdate
      parameters:
      - name: speciesId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeciesRequestPayload'
        required: true
      responses:
        "200":
          description: Species updated or merged with an existing species.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccessResponsePayload'
        "404":
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
    delete:
      tags:
      - GISApp
      summary: Delete an existing species.
      operationId: speciesDelete
      parameters:
      - name: speciesId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Species deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccessResponsePayload'
        "404":
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
        "409":
          description: Cannot delete the species because it is currently in use.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccessResponsePayload'
...
components:
  schemas:
...
    AllFieldValuesPayload:
      required:
      - partial
      - values
      type: object
      properties:
        values:
          type: array
          description: "All the values this field could possibly have, whether or\
            \ not any accessions have them. For fields that allow the user to enter\
            \ arbitrary values, this is equivalent to querying the list of values\
            \ without any filter criteria, that is, it's a list of all the user-entered\
            \ values."
          items:
            type: string
        partial:
          type: boolean
          description: "If true, the list of values is too long to return in its entirety\
            \ and \"values\" is a partial list."
    AndNodePayload:
      required:
      - children
      type: object
      description: "Search criterion that matches results that meet all of a set of\
        \ other search criteria. That is, if the list of children is x, y, and z,\
        \ this will require x AND y AND z."
      allOf:
      - $ref: '#/components/schemas/SearchNodePayload'
      - type: object
        properties:
          children:
            minItems: 1
            type: array
            items:
              $ref: '#/components/schemas/SearchNodePayload'
...
    CreateSpeciesRequestPayload:
      required:
      - name
      type: object
      properties:
        name:
          type: string
    CreateSpeciesResponsePayload:
      required:
      - details
      - status
      type: object
      properties:
        details:
          $ref: '#/components/schemas/SpeciesDetails'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    ErrorDetails:
      required:
      - message
      type: object
      properties:
        message:
          type: string
...
    FieldNodePayload:
      required:
      - field
      - type
      - values
      type: object
      allOf:
      - $ref: '#/components/schemas/SearchNodePayload'
      - type: object
        properties:
          field:
            $ref: '#/components/schemas/SearchField'
          values:
            minItems: 1
            type: array
            description: "List of values to match. For exact and fuzzy searches, a\
              \ list of at least one value to search for; the list may include null\
              \ to match accessions where the field does not have a value. For range\
              \ searches, the list must contain exactly two values, the minimum and\
              \ maximum; one of the values may be null to search for all values above\
              \ a minimum or below a maximum."
            items:
              type: string
              nullable: true
          type:
            type: string
            enum:
            - Exact
            - Fuzzy
            - Range
    FieldValuesPayload:
      required:
      - partial
      - values
      type: object
      properties:
        values:
          type: array
          description: "List of values in the matching accessions. If there are accessions\
            \ where the field has no value, this list will contain null (an actual\
            \ null value, not the string \"null\")."
          items:
            type: string
            nullable: true
        partial:
          type: boolean
          description: "If true, the list of values is too long to return in its entirety\
            \ and \"values\" is a partial list."
...
    ListAllFieldValuesRequestPayload:
      required:
      - fields
      type: object
      properties:
        fields:
          type: array
          items:
            $ref: '#/components/schemas/SearchField'
    ListAllFieldValuesResponsePayload:
      required:
      - results
      - status
      type: object
      properties:
        results:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AllFieldValuesPayload'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    ListFieldValuesRequestPayload:
      required:
      - fields
      type: object
      properties:
        fields:
          type: array
          items:
            $ref: '#/components/schemas/SearchField'
        filters:
          type: array
          items:
            $ref: '#/components/schemas/SearchFilter'
        search:
          oneOf:
          - $ref: '#/components/schemas/AndNodePayload'
          - $ref: '#/components/schemas/FieldNodePayload'
          - $ref: '#/components/schemas/NotNodePayload'
          - $ref: '#/components/schemas/OrNodePayload'
    ListFieldValuesResponsePayload:
      required:
      - results
      - status
      type: object
      properties:
        results:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/FieldValuesPayload'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    ListSpeciesResponsePayload:
      required:
      - status
      - values
      type: object
      properties:
        values:
          type: array
          items:
            $ref: '#/components/schemas/SpeciesDetails'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    NotNodePayload:
      required:
      - child
      type: object
      description: Search criterion that matches results that do not match a set of
        search criteria.
      allOf:
      - $ref: '#/components/schemas/SearchNodePayload'
      - type: object
        properties:
          child:
            $ref: '#/components/schemas/SearchNodePayload'
...
    OrNodePayload:
      type: object
      description: "Search criterion that matches results that meet any of a set of\
        \ other search criteria. That is, if the list of children is x, y, and z,\
        \ this will require x OR y OR z."
      allOf:
      - $ref: '#/components/schemas/SearchNodePayload'
...
    SearchField:
      type: string
      enum:
      - accessionNumber
      - active
      - bagNumber
      - collectedDate
      - collectionNotes
      - cutTestSeedsCompromised
      - cutTestSeedsEmpty
      - cutTestSeedsFilled
      - dryingEndDate
      - dryingMoveDate
      - dryingStartDate
      - endangered
      - estimatedSeedsIncoming
      - family
      - geolocation
      - germinationEndDate
      - germinationPercentGerminated
      - germinationSeedType
      - germinationSeedsGerminated
      - germinationSeedsSown
      - germinationStartDate
      - germinationSubstrate
      - germinationTestNotes
      - germinationTestType
      - germinationTreatment
      - landowner
      - latestGerminationTestDate
      - latestViabilityPercent
      - nurseryStartDate
      - primaryCollector
      - processingMethod
      - processingNotes
      - processingStartDate
      - rare
      - receivedDate
      - remainingGrams
      - remainingQuantity
      - remainingUnits
      - siteLocation
      - sourcePlantOrigin
      - species
      - state
      - storageCondition
      - storageLocation
      - storageNotes
      - storagePackets
      - storageStartDate
      - targetStorageCondition
      - totalGrams
      - totalQuantity
      - totalUnits
      - totalViabilityPercent
      - treesCollectedFrom
      - viabilityTestType
      - withdrawalDate
      - withdrawalDestination
      - withdrawalGrams
      - withdrawalNotes
      - withdrawalPurpose
      - withdrawalQuantity
      - withdrawalRemainingGrams
      - withdrawalRemainingQuantity
      - withdrawalRemainingUnits
      - withdrawalUnits
    SearchFilter:
      required:
      - field
      - type
      - values
      type: object
      properties:
        field:
          $ref: '#/components/schemas/SearchField'
        values:
          minLength: 1
          type: array
          description: "List of values to match. For exact and fuzzy searches, a list\
            \ of at least one value to search for; the list may include null to match\
            \ accessions where the field does not have a value. For range searches,\
            \ the list must contain exactly two values, the minimum and maximum; one\
            \ of the values may be null to search for all values above a minimum or\
            \ below a maximum."
          items:
            type: string
            nullable: true
        type:
          type: string
          enum:
          - Exact
          - Fuzzy
          - Range
    SearchNodePayload:
      required:
      - operation
      type: object
      properties:
        operation:
          type: string
      description: A search criterion. The search will return results that match this
        criterion. The criterion can be composed of other search criteria to form
        arbitrary Boolean search expressions.
      discriminator:
        propertyName: operation
        mapping:
          and: '#/components/schemas/AndNodePayload'
          field: '#/components/schemas/FieldNodePayload'
          not: '#/components/schemas/NotNodePayload'
          or: '#/components/schemas/NotNodePayload'
      oneOf:
      - $ref: '#/components/schemas/AndNodePayload'
      - $ref: '#/components/schemas/FieldNodePayload'
      - $ref: '#/components/schemas/NotNodePayload'
      - $ref: '#/components/schemas/OrNodePayload'
...
    SimpleErrorResponsePayload:
      required:
      - error
      - status
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetails'
        status:
          $ref: '#/components/schemas/SuccessOrError'
    SimpleSuccessResponsePayload:
      required:
      - status
      type: object
      properties:
        status:
          $ref: '#/components/schemas/SuccessOrError'
    SpeciesCreateResponsePayload:
      required:
      - id
      - status
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/SuccessOrError'
    SpeciesDetails:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    SpeciesGetResponsePayload:
      required:
      - species
      - status
      type: object
      properties:
        species:
          $ref: '#/components/schemas/SpeciesResponseElement'
        status:
          $ref: '#/components/schemas/SuccessOrError'
    SpeciesListResponsePayload:
      required:
      - species
      - status
      type: object
      properties:
        species:
          type: array
          items:
            $ref: '#/components/schemas/SpeciesResponseElement'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    SpeciesRequestPayload:
      required:
      - name
      type: object
      properties:
        conservationStatus:
          type: string
        familyId:
          type: integer
          format: int64
        isScientific:
          type: boolean
          description: True if name is the scientific name for the species.
        name:
          type: string
        plantForm:
          type: string
          enum:
          - Tree
          - Shrub
          - Vine
          - Liana
          - Herbaceous
        rare:
          type: string
          enum:
          - "No"
          - "Yes"
          - Unsure
        tsn:
          type: string
          description: Taxonomic serial number from ITIS database.
          externalDocs:
            url: https://en.wikipedia.org/wiki/Integrated_Taxonomic_Information_System
    SpeciesResponseElement:
      required:
      - id
      - isScientific
      - name
      type: object
      properties:
        conservationStatus:
          type: string
        familyId:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        isScientific:
          type: boolean
          description: True if name is the scientific name for the species.
        name:
          type: string
        plantForm:
          type: string
          enum:
          - Tree
          - Shrub
          - Vine
          - Liana
          - Herbaceous
        rare:
          type: string
          enum:
          - "No"
          - "Yes"
          - Unsure
        tsn:
          type: string
          description: Taxonomic serial number from ITIS database.
          externalDocs:
            url: https://en.wikipedia.org/wiki/Integrated_Taxonomic_Information_System
...
    SuccessOrError:
      type: string
      description: Indicates of success or failure of the requested operation.
      enum:
      - ok
      - error
...
    UpdateSpeciesResponsePayload:
      required:
      - status
      type: object
      properties:
        mergedWithSpeciesId:
          type: integer
          description: "If the requested species name already existed, the ID of the\
            \ existing species. Will not be present if the requested species name\
            \ did not already exist."
          format: int64
        status:
          $ref: '#/components/schemas/SuccessOrError'
...